\name{EColiNetwork}
\docType{data}
\alias{EColiNetwork}
\title{Example TF-target graph from Regulon DB.}
\description{
  This list contains the TF-target graph used in the vignette.
}
\usage{EColiNetwork}
\format{A list containing the target gene sets of 160 TFs.}
\value{A list containing the target gene sets of 160 TFs.}
\source{This TF-target graph was taken from (R. Castelo and A. Roverato, 2009). It is a pre-filtered version of RegulonDB 6.1.}
\references{
R. Castelo and A. Roverato. Reverse engineering molecular regulatory networks from microarray data with qp-graphs. \emph{J Comput Biol}, 16(2):213227, Feb 2009.
}
\keyword{datasets}
