\name{EColiOxygen}
\docType{data}
\alias{EColiOxygen}
\title{Example data set from E. Coli to sample TF activities.}
\description{
This data set gives expression values for three experiments of the E. Coli K12 strain under aerobic and three experiments under anaerobic growth. It is used in the vignette to illustrate application of birta to TFs only.
}
\usage{EColiOxygen}
\format{ExpressionSet}
\value{ExpressionSet}
\source{The original data comes from (Covert et al., 2004) The normalized data set used here is taken from the qpgraph package by R. Castelo and A. Roverato.}
\references{
M. W. Covert, E. M. Knight, J. L. Reed, M. J. Herrgard, and B. O. Palsson. Integrating high-throughput and computational data elucidates bacterial networks. \emph{Nature}, 429(6987):9296,May 2004.
}
\keyword{datasets}
