\name{plotConvergence}
\alias{plotConvergence}
\title{Plot the marginal log-likelihood of the model along MCMC samples (after thinning).}
\description{
Plot the marginal log-likelihood of the model along MCMC samples (after thinning).
}
\usage{
plotConvergence(res, title="")
}
\arguments{
\item{res}{The output of birte.run (a list object).}
\item{title}{Optional title of the plot.}
}
\value{
none
}
\seealso{\code{\link{birteRun}}}
\examples{
# see birteRun
}
\author{
Holger Froehlich
}