\name{TFexpr}
\docType{data}
\alias{TFexpr}
\title{Transcription factor expression values for the aerobic-anaerobic growth experiment.}
\description{
  This data set gives expression values for the 160 TF of the TF-target graph EColiNetwork used in the vignette. 
}
\usage{TFexpr}
\format{ExpressionSet. Rownames in the assayData correspond to entries in TF-target graph.}
\value{ExpressionSet. Rownames in the assayData correspond to entries in TF-target graph.}
\source{See EColiOxygen and EColiNetwork (see reference for details).}
\references{
M. W. Covert, E. M. Knight, J. L. Reed, M. J. Herrgard, and B. O. Palsson. Integrating high- throughput and computational data elucidates bacterial networks. \emph{Nature}, 429(6987):92-96, May 2004.
}
\keyword{datasets}
